Adds ReplicatedWriteLogInternal.h
Hopefully trims enough off of ReplicatedWriteLog.cc that gitlab will
show the diffs in a merge request.
